Log IP/device changes within the same session
Store IP and device information in the session and log when
it changes. The goal is to detect session leakage when the
session is accidentally sent to another user, which is a
hypothetical cause of T264370. The log will be noisy since
users do change IP addresses for a number of reasons,
but we are mainly interested in the ability of correlating
user-reported incidents where we have a username to filter
by, so that's OK.

Based on I27468a3f6d58.

Bug: T264799
Change-Id: Ifa14fa637c1b199159ea11e983a25212ae005565
(cherry picked from commit d5d3c90152299abf73f7747d5c53984d9fb53ea1)