media: Support GPSAltitudeRef exif tag
JPG images generated by the Nikon COOLPIX AW100 camera (and others)
create bogus non-numeric GPSAltitudeRef tags in their EXIF data. That
may not be the camera's fault, maybe it's actually PHP which is
misparsing the tag -- the value is the "unicode replacement character"
U+FFFD which is very suspicious.  In any case, we have a decent number
of images on commons uploaded from these cameras as it turns out, so
it generates a large-ish amount of logspam if we can't handle their
bogus data correctly. Add support for GPSAltitudeRef, including the
non-standard non-numeric value emitted when photos from these cameras
are processed.

Bug: T267370
Change-Id: If85081f023a59b56557b9a74b8431c7be332d7e5